Dedicated fn to load dataset and async watcher
